Patch mousePressed library function
`p5.Element.prototype.mousePressed` has a decorator function wrapping
the user's callback `fxn`. When this decorator calls the user's `fxn` is
never passes the `MouseEvent event` through. I may fork p5js.git and
submit a pull request, but it is literally a single word difference. I
commented on the issue which introduced this bug, and I'll see where it
goes from there.